Just a correction on searching graph function
When searching for the graphs we always want to search by both search terms. By having that if else statement made it so we always search by a single term which made the searching inconsistent.

Got rid of it, now we can search by the queries by the cardio and gene type